updated code on DCC for internal consistency

+++ fetch-gbk-annotation.py
@@ -10,7 +10,7 @@ from the src/ directory as python fetch-gbk-annotation.py.
 '''
 
 import urllib.request
-from os.path import basename
+from os.path import basename, exists
 import gzip
 import os
 from tqdm import tqdm
@@ -24,15 +24,16 @@ with open("../results/prokaryotes-with-plasmids.txt", "r") as genome_report_fh:
             continue ## don't process the header further.
         fields = line.split('\t')
         ftp_path = fields[20]
-        ## Now download the Genbank annotation.
+        ## Now download the Genbank annotation if it doesn't exist on disk.
         gbk_ftp_path = ftp_path + '/' + basename(ftp_path) + "_genomic.gbff.gz"
         gbff_gz_fname = "../results/gbk-annotation/" + basename(ftp_path) + "_genomic.gbff.gz"
-        try:
-            urllib.request.urlretrieve(gbk_ftp_path, filename=gbff_gz_fname)
-        except urllib.error.URLError: ## some problem happens
-            try: ## try one more time, in case there was some connection problem.
-                urllib.request.urlretrieve(gbk_ftp_path,filename=gbff_gz_fname)
-            except: ## skip if the translated cds don't exist.
-                continue
+        if not exists(gbff_gz_fname):
+            try:
+                urllib.request.urlretrieve(gbk_ftp_path, filename=gbff_gz_fname)
+            except urllib.error.URLError: ## some problem happens
+                try: ## try one more time, in case there was some connection problem.
+                    urllib.request.urlretrieve(gbk_ftp_path,filename=gbff_gz_fname)
+                except: ## skip if still can't get it.
+                    continue
